add Jwt config roles

+++ src/main/java/com/fabricaescuela20241/reserva/Servicio/ServicioReserva.java
@@ -1,6 +1,7 @@
 package com.fabricaescuela20241.reserva.Servicio;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,10 +17,10 @@ public class ServicioReserva {
 
     public Reserva guardarReserva(Reserva reserva) {
         return repositorioReserva.save(reserva);
+
     }
 
     public List<Reserva> obtenerTodasLasReservas() {
-        System.out.println("Se está obteniendo la lista de todas las reservas");
         return repositorioReserva.findAll();
     }
 
@@ -27,13 +28,45 @@ public class ServicioReserva {
         return repositorioReserva.findById(id).orElse(null);
     }
 
-    public void eliminarReservaPorId(int id) {
-        System.out.println("Se está eliminando la reserva");
+    public boolean eliminarReservaPorId(int id) {
+        if(!repositorioReserva.existsById(id)){
+            return false;
+        }
         repositorioReserva.deleteById(id);
+        return true;
     }
 
-    public long contarReservas(){
-        return repositorioReserva.count();
+    public boolean actualizarReserva(int id, Reserva reserva){
+        Optional<Reserva> optionalReserva = repositorioReserva.findById(id);
+        if(!optionalReserva.isPresent()){
+            return false;
+        }
+
+        Reserva reservaExiste = optionalReserva.get();
+
+        if (reserva.getIdReserva() != id){
+            return false;
+        }
+
+        if(reserva.getFechaReserva() != null){
+            reservaExiste.setFechaReserva(reserva.getFechaReserva());
+        }
+
+        if (reserva.getEstado() != null){
+            reservaExiste.setEstado(reserva.getEstado());
+        }
+
+        if (reserva.getVuelo() != null){
+            reservaExiste.setVuelo(reserva.getVuelo());
+        }
+
+        if (reserva.getPrecio() != null){
+            reservaExiste.setPrecio(reserva.getPrecio());
+        }
+
+        repositorioReserva.save(reservaExiste);
+        return true;
+
     }
 
 }
